+++ src/CHANGELOG.md
@@ -4,6 +4,8 @@ OpenBLCMM Changelog
 **v1.4.1** *(unreleased)*
 - Fixed a crash which could happen when hitting next/prev quickly when searching
   in Object Explorer dumps or code edit window.
+- Fixed "Flatten category contents" action to always leave statements in the
+  same order they started in.
 - Various updates to double-click-to-select and ctrl-shift-left/right selection
   mechanics:
   - Both were nearly completely rewritten with a better method of overwriting